split pyvector_integrator to several files
