+++ petram/helper/integrators/pyvectorfe_integrator_base.py
@@ -78,67 +78,3 @@ class PyVectorFEIntegratorBase(mfem.PyBilinearFormIntegrator):
 
         return np.stack(ret) # shape sdim (dir. grad), Dof, sdim
 
-class PyVectorFEPartialIntegrator(PyVectorIntegratorBase):
-    '''
-  
-    vec(v)_i lamblda(i, j, k)  \partial_j vec(u)_k
-
-    V and U are ND
-
-    '''
-    def __init__(self, lam, ir=None):
-        PyVectorIntegratorBase.__init__(self, ir)
-
-        self._ir = self.GetIntegrationRule()
-
-        self.tr_vshape = mfem.DenseMatrix()
-        self.te_vshape = mfem.DenseMatrix()
-
-        self.partelmat = mfem.DenseMatrix()
-        self.val = mfem.Vector()
-
-
-    def AssembleElementMatrix(self, el, trans, elmat):
-        self.AssembleElementMatrix2(el, el, trans, elmat)
-
-    def AssembleElementMatrix2(self, trial_fe, test_fe, trans, elmat):
-        if self._ir is None:
-            self.set_ir(trial_fe, test_fe, trans)
-
-        sdim = trans.GetSpaceDim()
-        tr_vdim = max(sdim, trial_fe.GetRangeDim())
-        te_vdim = max(sdim, test_fe.GetRangeDim())
-
-        tr_nd = trial_fe.GetDof()
-        te_nd = test_fe.GetDof()
-        
-        self.partelmat.SetSize(te_nd, tr_nd)        
-
-        self.te_vshape.SetSize(te_nd, te_vdim)
-        self.val.SetSize(tr_vdim**3)
-        
-        itrans = mfem.InverseElementTransformation(trans)
-        
-        elmat.Assign(0.0)
-        
-        ds = self.get_ds(self.ir, trans)
-        rr = range(tr_vdim)
-        
-        for i in range(ir.GetNPoints()):
-            
-            ip = ir.IntPoint(i)
-            test_fe.CalcVshape(trans, te_vshape)
-
-            # test_i lamblda(i, j, k)  \partial_j trial_k
-            self.lam.Eval(self.val, trans, ip)
-            lam = self.val.GetDataArray().reshape(tr_vdi, tr_vdim, tr_vdim)
-            
-            dshape = self.get_dshape(trial_fe, ip, trans, itrans, ds)
-
-            # DoF sdim * sdim(grad) DoF sdim
-            udvdx = np.tensordot(te_vshape.GetDataArray(),
-                                 dshape)* ip.weight * trans.Weight()
-            
-
-            for ii, jj, kk in prod(rr, rr, rr):
-                elmat[:,:] += lam[ii, jj, kk]*udvdx[:, ii, jj, :, kk]
